+++ tests/e2e-playwright/tests/jupyterlab.py
@@ -11,6 +11,7 @@ from typing import Final
 
 from playwright.sync_api import APIRequestContext, Page
 from pydantic import AnyUrl
+from pytest_simcore.playwright_utils import on_web_socket, test_logger
 from tenacity import Retrying
 from tenacity.retry import retry_if_exception_type
 from tenacity.stop import stop_after_attempt
@@ -19,13 +20,6 @@ from tenacity.wait import wait_fixed
 projects_uuid_pattern: Final[re.Pattern] = re.compile(
     r"/projects/([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})"
 )
-
-
-def on_web_socket(ws) -> None:
-    print(f"WebSocket opened: {ws.url}")
-    ws.on("framesent", lambda payload: print("⬇️", payload))
-    ws.on("framereceived", lambda payload: print("⬆️", payload))
-    ws.on("close", lambda payload: print("WebSocket closed"))
 
 
 def test_jupyterlab(
@@ -56,7 +50,7 @@ def test_jupyterlab(
         page.wait_for_timeout(1000)
 
     # Get project uuid, will be used to delete this project in the end
-    print(f"projects uuid endpoint captured: {response_info.value.url}")
+    test_logger.info(f"projects uuid endpoint captured: {response_info.value.url}")
     match = projects_uuid_pattern.search(response_info.value.url)
     assert match
     extracted_uuid = match.group(1)
@@ -84,7 +78,8 @@ def test_jupyterlab(
         _jupyterlab_ui.fill("print('test')")
         _jupyterlab_ui.press("Shift+Enter")
     else:
-        raise ValueError("Not supported service key")
+        msg = "Not supported service key"
+        raise ValueError(msg)
     page.wait_for_timeout(1000)
 
     # Going back to dashboard
