+++ tests/e2e-playwright/tests/sim4life.py
@@ -5,15 +5,13 @@
 # pylint: disable=too-many-statements
 # pylint: disable=unnecessary-lambda
 
-import logging
 import re
-from contextlib import ExitStack
 from http import HTTPStatus
 from typing import Final
 
 from playwright.sync_api import APIRequestContext, Page
 from pydantic import AnyUrl
-from pytest_simcore.playwright_utils import log_context, test_logger
+from pytest_simcore.playwright_utils import on_web_socket, test_logger
 from tenacity import Retrying
 from tenacity.retry import retry_if_exception_type
 from tenacity.stop import stop_after_attempt
@@ -22,23 +20,6 @@ from tenacity.wait import wait_fixed
 projects_uuid_pattern: Final[re.Pattern] = re.compile(
     r"/projects/([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})"
 )
-
-
-def on_web_socket(ws) -> None:
-    stack = ExitStack()
-    ctx = stack.enter_context(
-        log_context(
-            logging.INFO,
-            (
-                f"WebSocket opened: {ws.url}",
-                "WebSocket closed",
-            ),
-        )
-    )
-
-    ws.on("framesent", lambda payload: ctx.logger.info(payload))
-    ws.on("framereceived", lambda payload: ctx.logger.info(payload))
-    ws.on("close", lambda payload: stack.close())
 
 
 def test_sim4life(
